freebayes filter and normalise resumable scripts

+++ popgen/freebayes/run_filter.sh
@@ -33,7 +33,9 @@ SEPARATOR=" "
 VCFTOOLS_ARG="$( printf "${SEPARATOR}%s" "${VCFFILES[@]}" )"
 
 # > STEP 3: Merge individual VCFs
-bcftools merge -Ov -o btrys06.merged.vcf ${VCFTOOLS_ARG}
+if [[ ! -f ${PREFIX}.merged.vcf  ]]; then
+    bcftools merge -Ov -o ${PREFIX}.merged.vcf ${VCFTOOLS_ARG}
+fi
 
 # > STEP 3: Filter vcfs according to publication Pete emailed to pro_zac
 MISSING=0.5 ## this means >50% of individuals need to have the site
@@ -41,13 +43,21 @@ MINQ=30 ## minimum SNP quality of 30, whatever that means
 MAC=1 ## minor allele count must be >= 1
 MINDEPTH=3 ## minimum depth of 3 for a genotype call
 MAF=0.05 ## minor allele frequency greater than or equal to 0.05
-vcftools --vcf ${PREFIX}.merged.vcf --max-missing ${MISSING} --mac ${MAC} --minQ ${MINQ} --min-meanDP ${MINDEPTH} --remove-filtered-all --recode --recode-INFO-all --maf ${MAF} --out ${PREFIX}.filtered.vcf
-mv ${PREFIX}.filtered.vcf.recode.vcf ${PREFIX}.filtered.vcf
+if [[ ! -f ${PREFIX}.filtered.vcf  ]]; then
+    vcftools --vcf ${PREFIX}.merged.vcf --max-missing ${MISSING} --mac ${MAC} --minQ ${MINQ} --min-meanDP ${MINDEPTH} --remove-filtered-all --recode --recode-INFO-all --maf ${MAF} --out ${PREFIX}.filtered.vcf;
+    mv ${PREFIX}.filtered.vcf.recode.vcf ${PREFIX}.filtered.vcf;
+fi
 
 # > STEP 4: Remove indels
-bcftools view --exclude-types indels -Ov -o ${PREFIX}.filtered.noindels.vcf ${PREFIX}.filtered.vcf
+if [[ ! -f ${PREFIX}.filtered.noindels.vcf  ]]; then
+    bcftools view --exclude-types indels -Ov -o ${PREFIX}.filtered.noindels.vcf ${PREFIX}.filtered.vcf
+fi
 
 # > STEP 5: More custom filtering
 POPMISSING=0.5 ## remove a site if each population does not have at least this much presence
-python ${VARSCRIPTDIR}/popgen/VCF/filter_vcf.py -v ${PREFIX}.filtered.noindels.vcf -p ${POPSFILE} -o ${PREFIX}.final.vcf --mpp ${POPMISSING}
-python ${VARSCRIPTDIR}/popgen/VCF/filter_vcf.py -v ${PREFIX}.filtered.noindels.vcf -p ${POPSFILE} --geno -o ${PREFIX}.final.geno --mpp ${POPMISSING}
+if [[ ! -f ${PREFIX}.final.vcf  ]]; then
+    python ${VARSCRIPTDIR}/popgen/VCF/filter_vcf.py -v ${PREFIX}.filtered.noindels.vcf -p ${POPSFILE} -o ${PREFIX}.final.vcf --mpp ${POPMISSING}
+fi
+if [[ ! -f ${PREFIX}.final.geno  ]]; then
+    python ${VARSCRIPTDIR}/popgen/VCF/filter_vcf.py -v ${PREFIX}.filtered.noindels.vcf -p ${POPSFILE} --geno -o ${PREFIX}.final.geno --mpp ${POPMISSING}
+fi
